+++ src/core/server/commands/debugVehicle.ts
@@ -7,19 +7,27 @@ import { playerFuncs } from '../extensions/Player';
 import VehicleFuncs from '../extensions/VehicleFuncs';
 import { LocaleController } from '../../shared/locale/locale';
 import { LOCALE_KEYS } from '../../shared/locale/languages/keys';
+import { VEHICLE_OWNERSHIP } from '../../shared/flags/VehicleOwnershipFlags';
 
 ChatController.addCommand(
     'vehicle',
     LocaleController.get(LOCALE_KEYS.COMMAND_VEHICLE, '/vehicle'),
     PERMISSIONS.ADMIN,
-    handleTemp
+    handleTemp,
 );
 
 ChatController.addCommand(
     'addvehicle',
     LocaleController.get(LOCALE_KEYS.COMMAND_ADD_VEHICLE, '/addvehicle'),
     PERMISSIONS.ADMIN,
-    handleAdd
+    handleAdd,
+);
+
+ChatController.addCommand(
+    'addfactionvehicle',
+    '/addfactionvehicle [model]',
+    PERMISSIONS.ADMIN,
+    handleAddFactionVehicle,
 );
 
 function handleTemp(player: alt.Player, model: string): void {
@@ -60,8 +68,49 @@ function handleAdd(player: alt.Player, model: string): void {
 
         VehicleFuncs.add(
             { owner: player.data._id.toString(), fuel: 100, model, position: veh.pos, rotation: veh.rot },
-            false
+            false,
         );
+        veh.destroy();
+    } catch (err) {
+        console.log(err);
+        playerFuncs.emit.message(player, LocaleController.get(LOCALE_KEYS.INVALID_VEHICLE_MODEL));
+        return;
+    }
+}
+
+function handleAddFactionVehicle(player: alt.Player, model: string): void {
+    if (!player.data || !player.data.faction) {
+        playerFuncs.emit.message(player, `Must be in a faction to add that vehicle to the faction.`);
+        return;
+    }
+
+    if (!model) {
+        playerFuncs.emit.message(player, `Vehicle model not specified`);
+        return;
+    }
+
+    if (player.data.isDead) {
+        playerFuncs.emit.message(player, LocaleController.get(LOCALE_KEYS.CANNOT_PERFORM_WHILE_DEAD));
+        return;
+    }
+
+    const fwd = getVectorInFrontOfPlayer(player, 5);
+
+    try {
+        const veh = VehicleFuncs.tempVehicle(player, model, fwd, new alt.Vector3(0, 0, 0));
+
+        VehicleFuncs.add(
+            {
+                owner: player.data.faction.toString(),
+                ownerType: VEHICLE_OWNERSHIP.FACTION,
+                fuel: 100,
+                model,
+                position: veh.pos,
+                rotation: veh.rot,
+            },
+            false,
+        );
+
         veh.destroy();
     } catch (err) {
         console.log(err);