+++ src/core/shared/flags/VehicleOwnershipFlags.ts
@@ -1,5 +1,3 @@
 export enum VEHICLE_OWNERSHIP {
-    NONE = 0,
-    PLAYER = 1,
-    FACTION = 2
+    FACTION = 1,
 }